requirements.txt
Django==4.2.5
django_filter==22.1
djangorestframework==3.14.0
drf_extra_fields==3.4.1
drf-spectacular==0.27.1
python-dotenv==0.21.1